+++ src/lib/pages/SecretTab.svelte
@@ -1,0 +1,11 @@
+<script>
+    // Secret
+    import Secret from '../components/Secret.svelte';   
+</script>
+
+<Secret/>
+
+<style>
+
+
+</style>