+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 You can update your OpenTaiko Hub version directly from the releases tab:
 - https://github.com/OpenTaiko/OpenTaiko-Hub/releases
 
+## [0.1.6] - 2025-07-?? (Alpha)
+
+- [Feat] Added the current version number to the OpenTaiko Hub.
+- [Feat] Added a update system for the OpenTaiko Hub.
+- [Feat] Added the "Secret" tab. (Maybe go take a look...)
+- [Feat] Added the "Links" tab.
+- [Feat] Added the "Lyrics" tab under "Tools".
+- [Feat] Added Subtitle Edit to the "Lyrics" tab under "Tools".
+- [Feat] Added contributor credits for the OpenTaiko Hub to the "Credits" tab under "Information".
+- [Enhancement] Some UI changes.
+- [Chore] Changed the GitHub link to go to the OpenTaiko Hub's repository.
+- [Chore] Moved the main OpenTaiko repository link to the "Links" tab.
+- [Chore] Moved the Discord link to the "Links" tab.
+- [Chore] Updated the HoF crowns to link to https://opentaiko.github.io/.
+- [Chore] Fixed some typos.
+
 ## [0.1.5] - 2025-05-02 (Alpha)
 
 - [Fix] AppImage should now be fully functional on Linux